Moved Mesh to a separate file.
